Add FeaturesCardView component and integrate it into NewBookSection; remove deprecated commitment section from Footer
